Refactoring Dataobject Grid Listing Setup in Backend (#3796)
