+++ bundles/AdminBundle/Controller/Admin/DataObject/VariantsController.php
@@ -15,6 +15,7 @@
 namespace Pimcore\Bundle\AdminBundle\Controller\Admin\DataObject;
 
 use Pimcore\Bundle\AdminBundle\Controller\AdminController;
+use Pimcore\Bundle\AdminBundle\Helper\GridHelperService;
 use Pimcore\Model\DataObject;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
@@ -56,12 +57,13 @@ class VariantsController extends AdminController
      * @Route("/get-variants", methods={"GET", "POST"})
      *
      * @param Request $request
+     * @param GridHelperService $gridHelperService
      *
      * @return JsonResponse
      *
      * @throws \Exception
      */
-    public function getVariantsAction(Request $request)
+    public function getVariantsAction(Request $request, GridHelperService $gridHelperService)
     {
         // get list of variants
 
@@ -122,83 +124,15 @@ class VariantsController extends AdminController
             }
 
             if ($parentObject->isAllowed('view')) {
-                $class = $parentObject->getClass();
-                $className = $parentObject->getClass()->getName();
+                $allParams = array_merge($request->request->all(), $request->query->all());
 
-                $start = 0;
-                $limit = 15;
-                $orderKey = 'oo_id';
-                $order = 'ASC';
+                //specify a few special params
+                $allParams['folderId'] = $parentObject->getId();
+                $allParams['only_direct_children'] == 'true';
+                $allParams['classId'] = $parentObject->getClassId();
 
-                $fields = [];
-                $bricks = [];
-                if ($request->get('fields')) {
-                    $fields = $request->get('fields');
+                $list = $gridHelperService->prepareListingForGrid($allParams, $request->getLocale(), $this->getAdminUser());
 
-                    foreach ($fields as $f) {
-                        $parts = explode('~', $f);
-                        if (substr($f, 0, 1) == '~') {
-                            $type = $parts[1];
-                        } elseif (count($parts) > 1) {
-                            $bricks[$parts[0]] = $parts[0];
-                        }
-                    }
-                }
-
-                if ($request->get('limit')) {
-                    $limit = $request->get('limit');
-                }
-                if ($request->get('start')) {
-                    $start = $request->get('start');
-                }
-
-                $orderKey = 'o_id';
-                $order = 'ASC';
-
-                $colMappings = [
-                    'filename' => 'o_key',
-                    'fullpath' => ['o_path', 'o_key'],
-                    'id' => 'oo_id',
-                    'published' => 'o_published',
-                    'modificationDate' => 'o_modificationDate',
-                    'creationDate' => 'o_creationDate'
-                ];
-
-                $sortingSettings = \Pimcore\Bundle\AdminBundle\Helper\QueryParams::extractSortingSettings(array_merge($request->request->all(), $request->query->all()));
-                if ($sortingSettings['orderKey'] && $sortingSettings['order']) {
-                    $orderKey = $sortingSettings['orderKey'];
-                    if (array_key_exists($orderKey, $colMappings)) {
-                        $orderKey = $colMappings[$orderKey];
-                    }
-                    $order = $sortingSettings['order'];
-                }
-
-                if ($request->get('dir')) {
-                    $order = $request->get('dir');
-                }
-
-                $listClass = '\\Pimcore\\Model\\DataObject\\' . ucfirst($className) . '\\Listing';
-
-                $conditionFilters = ['o_parentId = ' . $parentObject->getId()];
-                // create filter condition
-                if ($request->get('filter')) {
-                    $conditionFilters[] = DataObject\Service::getFilterCondition($request->get('filter'), $class);
-                }
-                if ($request->get('condition')) {
-                    $conditionFilters[] = '(' . $request->get('condition') . ')';
-                }
-
-                $list = new $listClass();
-                if (!empty($bricks)) {
-                    foreach ($bricks as $b) {
-                        $list->addObjectbrick($b);
-                    }
-                }
-                $list->setCondition(implode(' AND ', $conditionFilters));
-                $list->setLimit($limit);
-                $list->setOffset($start);
-                $list->setOrder($order);
-                $list->setOrderKey($orderKey);
                 $list->setObjectTypes([DataObject\AbstractObject::OBJECT_TYPE_VARIANT]);
 
                 $list->load();
@@ -206,7 +140,7 @@ class VariantsController extends AdminController
                 $objects = [];
                 foreach ($list->getObjects() as $object) {
                     if ($object->isAllowed('view')) {
-                        $o = DataObject\Service::gridObjectData($object, $fields);
+                        $o = DataObject\Service::gridObjectData($object, $request->get('fields'));
                         $objects[] = $o;
                     }
                 }
