+++ models/DataObject/Service.php
@@ -719,270 +719,6 @@ class Service extends Model\Element\Service
     }
 
     /**
-     *
-     * @param string $filterJson
-     * @param ClassDefinition $class
-     * @param $requestedLanguage
-     *
-     * @return string
-     */
-    public static function getFeatureFilters($filterJson, $class, $requestedLanguage)
-    {
-        $joins = [];
-        $conditions = [];
-
-        if ($filterJson) {
-            $filters = json_decode($filterJson, true);
-            foreach ($filters as $filter) {
-                $operator = '=';
-
-                $filterField = $filter['property'];
-                $filterOperator = $filter['operator'];
-
-                if ($filter['type'] == 'string') {
-                    $operator = 'LIKE';
-                } elseif ($filter['type'] == 'numeric') {
-                    if ($filterOperator == 'lt') {
-                        $operator = '<';
-                    } elseif ($filterOperator == 'gt') {
-                        $operator = '>';
-                    } elseif ($filterOperator == 'eq') {
-                        $operator = '=';
-                    }
-                } elseif ($filter['type'] == 'date') {
-                    if ($filterOperator == 'lt') {
-                        $operator = '<';
-                    } elseif ($filterOperator == 'gt') {
-                        $operator = '>';
-                    } elseif ($filterOperator == 'eq') {
-                        $operator = '=';
-                    }
-                    $filter['value'] = strtotime($filter['value']);
-                } elseif ($filter['type'] == 'list') {
-                    $operator = '=';
-                } elseif ($filter['type'] == 'boolean') {
-                    $operator = '=';
-                    $filter['value'] = (int)$filter['value'];
-                }
-
-                $keyParts = explode('~', $filterField);
-
-                if (substr($filterField, 0, 1) != '~') {
-                    continue;
-                }
-
-                $type = $keyParts[1];
-                if ($type != 'classificationstore') {
-                    continue;
-                }
-
-                $fieldName = $keyParts[2];
-                $groupKeyId = explode('-', $keyParts[3]);
-
-                /** @var $csFieldDefinition Model\DataObject\ClassDefinition\Data\Classificationstore */
-                $csFieldDefinition = $class->getFieldDefinition($fieldName);
-
-                $language = $requestedLanguage;
-                if (!$csFieldDefinition->isLocalized()) {
-                    $language = 'default';
-                }
-
-                $groupId = $groupKeyId[0];
-                $keyid = $groupKeyId[1];
-
-                $keyConfig = Model\DataObject\Classificationstore\KeyConfig::getById($keyid);
-                $type = $keyConfig->getType();
-                $definition = json_decode($keyConfig->getDefinition());
-                $field = \Pimcore\Model\DataObject\Classificationstore\Service::getFieldDefinitionFromJson($definition, $type);
-
-                if ($field instanceof ClassDefinition\Data) {
-                    $mappedKey = 'cskey_' . $fieldName . '_' . $groupId . '_' . $keyid;
-                    $joins[] = ['fieldname' => $fieldName, 'groupId' => $groupId, 'keyId' => $keyid, 'language' => $language];
-                    $condition = $field->getFilterConditionExt(
-                        $filter['value'],
-                        $operator,
-                        [
-                            'name' => $mappedKey]
-                    );
-
-                    $conditions[$mappedKey] = $condition;
-                }
-            }
-        }
-
-        $result = [
-            'joins' => $joins,
-            'conditions' => $conditions
-        ];
-
-        return $result;
-    }
-
-    /**
-     *
-     * @param string $filterJson
-     * @param ClassDefinition $class
-     *
-     * @return string
-     */
-    public static function getFilterCondition($filterJson, $class)
-    {
-        $systemFields = self::getSystemFields();
-
-        // create filter condition
-        $conditionPartsFilters = [];
-
-        if ($filterJson) {
-            $db = \Pimcore\Db::get();
-            $filters = json_decode($filterJson, true);
-            foreach ($filters as $filter) {
-                $operator = '=';
-
-                $filterField = $filter['property'];
-                $filterOperator = $filter['operator'];
-
-                if ($filter['type'] == 'string') {
-                    $operator = 'LIKE';
-                } elseif ($filter['type'] == 'numeric') {
-                    if ($filterOperator == 'lt') {
-                        $operator = '<';
-                    } elseif ($filterOperator == 'gt') {
-                        $operator = '>';
-                    } elseif ($filterOperator == 'eq') {
-                        $operator = '=';
-                    }
-                } elseif ($filter['type'] == 'date') {
-                    if ($filterOperator == 'lt') {
-                        $operator = '<';
-                    } elseif ($filterOperator == 'gt') {
-                        $operator = '>';
-                    } elseif ($filterOperator == 'eq') {
-                        $operator = '=';
-                    }
-                    $filter['value'] = strtotime($filter['value']);
-                } elseif ($filter['type'] == 'list') {
-                    $operator = '=';
-                } elseif ($filter['type'] == 'boolean') {
-                    $operator = '=';
-                    $filter['value'] = (int)$filter['value'];
-                }
-
-                $field = $class->getFieldDefinition($filterField);
-                $brickField = null;
-                $brickType = null;
-                if (!$field) {
-
-                    // if the definition doesn't exist check for a localized field
-                    $localized = $class->getFieldDefinition('localizedfields');
-                    if ($localized instanceof ClassDefinition\Data\Localizedfields) {
-                        $field = $localized->getFieldDefinition($filterField);
-                    }
-
-                    //if the definition doesn't exist check for object brick
-                    $keyParts = explode('~', $filterField);
-
-                    if (substr($filterField, 0, 1) == '~') {
-                        // not needed for now
-//                            $type = $keyParts[1];
-//                            $field = $keyParts[2];
-//                            $keyid = $keyParts[3];
-                    } elseif (count($keyParts) > 1) {
-                        $brickType = $keyParts[0];
-                        $brickKey = $keyParts[1];
-
-                        $key = self::getFieldForBrickType($class, $brickType);
-                        $field = $class->getFieldDefinition($key);
-
-                        if (strpos($brickType, '?') !== false) {
-                            $brickDescriptor = substr($brickType, 1);
-                            $brickDescriptor = json_decode($brickDescriptor, true);
-                            $brickType = $brickDescriptor['containerKey'];
-                        }
-
-                        $brickClass = Objectbrick\Definition::getByKey($brickType);
-
-                        if ($brickDescriptor) {
-                            $brickField = $brickClass->getFieldDefinition('localizedfields')->getFieldDefinition($brickDescriptor['brickfield']);
-                        } else {
-                            $brickField = $brickClass->getFieldDefinition($brickKey);
-                        }
-                    }
-                }
-                if ($field instanceof ClassDefinition\Data\Objectbricks || $brickDescriptor) {
-                    // custom field
-                    $db = \Pimcore\Db::get();
-                    $brickPrefix = '';
-
-                    $ommitPrefix = false;
-
-                    if ($brickField instanceof Model\DataObject\ClassDefinition\Data\Checkbox
-                        || (($brickField instanceof Model\DataObject\ClassDefinition\Data\Date || $brickField instanceof Model\DataObject\ClassDefinition\Data\Datetime) && $brickField->getColumnType() == 'datetime')
-                    ) {
-                        $ommitPrefix = true;
-                    }
-
-                    if (!$ommitPrefix) {
-                        if ($brickDescriptor) {
-                            $brickPrefix = $db->quoteIdentifier($brickType . '_localized') . '.';
-                        } else {
-                            $brickPrefix = $db->quoteIdentifier($brickType) . '.';
-                        }
-                    }
-                    if (is_array($filter['value'])) {
-                        $fieldConditions = [];
-                        foreach ($filter['value'] as $filterValue) {
-                            $fieldConditions[] = $brickPrefix . $brickField->getFilterCondition($filterValue, $operator,
-                                    ['brickType' => $brickType]
-                                );
-                        }
-                        $conditionPartsFilters[] = '(' . implode(' OR ', $fieldConditions) . ')';
-                    } else {
-                        $conditionPartsFilters[] = $brickPrefix . $brickField->getFilterCondition($filter['value'], $operator,
-                                ['brickType' => $brickType]);
-                    }
-                } elseif ($field instanceof ClassDefinition\Data) {
-                    // custom field
-                    if (is_array($filter['value'])) {
-                        $fieldConditions = [];
-                        foreach ($filter['value'] as $filterValue) {
-                            $fieldConditions[] = $field->getFilterCondition($filterValue, $operator);
-                        }
-
-                        if (!empty($fieldConditions)) {
-                            $conditionPartsFilters[] = '(' . implode(' OR ', $fieldConditions) . ')';
-                        }
-                    } else {
-                        $conditionPartsFilters[] = $field->getFilterCondition($filter['value'], $operator);
-                    }
-                } elseif (in_array('o_' . $filterField, $systemFields)) {
-                    // system field
-                    if ($filterField == 'fullpath') {
-                        $conditionPartsFilters[] = 'concat(o_path, o_key) ' . $operator . ' ' . $db->quote('%' . $filter['value'] . '%');
-                    } elseif ($filterField == 'key') {
-                        $conditionPartsFilters[] = 'o_key ' . $operator . ' ' . $db->quote('%' . $filter['value'] . '%');
-                    } else {
-                        if ($filter['type'] == 'date' && $operator == '=') {
-                            //if the equal operator is chosen with the date type, condition has to be changed
-                            $maxTime = $filter['value'] + (86400 - 1); //specifies the top point of the range used in the condition
-                            $conditionPartsFilters[] = '`o_' . $filterField . '` BETWEEN ' . $db->quote($filter['value']) . ' AND ' . $db->quote($maxTime);
-                        } else {
-                            $conditionPartsFilters[] = '`o_' . $filterField . '` ' . $operator . ' ' . $db->quote($filter['value']);
-                        }
-                    }
-                }
-            }
-        }
-
-        $conditionFilters = '1 = 1';
-        if (count($conditionPartsFilters) > 0) {
-            $conditionFilters = '(' . implode(' AND ', $conditionPartsFilters) . ')';
-        }
-        Logger::log('DataObjectController filter condition:' . $conditionFilters);
-
-        return $conditionFilters;
-    }
-
-    /**
      * @static
      *
      * @param $object
@@ -1690,54 +1426,6 @@ class Service extends Model\Element\Service
         }
 
         return null;
-    }
-
-    /** Adds all the query stuff that is needed for displaying, filtering and exporting the feature grid data.
-     * @param $list list
-     * @param $featureJoins
-     * @param $class
-     * @param $featureFilters
-     */
-    public static function addGridFeatureJoins($list, $featureJoins, $class, $featureFilters)
-    {
-        if ($featureJoins) {
-            $me = $list;
-            $list->onCreateQuery(function (QueryBuilder $select) use ($list, $featureJoins, $class, $featureFilters, $me) {
-                $db = \Pimcore\Db::get();
-
-                $alreadyJoined = [];
-
-                foreach ($featureJoins as $featureJoin) {
-                    $fieldname = $featureJoin['fieldname'];
-                    $mappedKey = 'cskey_' . $fieldname . '_' . $featureJoin['groupId'] . '_' . $featureJoin['keyId'];
-                    if (isset($alreadyJoined[$mappedKey]) && $alreadyJoined[$mappedKey]) {
-                        continue;
-                    }
-                    $alreadyJoined[$mappedKey] = 1;
-
-                    $table = $me->getDao()->getTableName();
-                    $select->joinLeft(
-                        [$mappedKey => 'object_classificationstore_data_' . $class->getId()],
-                        '('
-                        . $mappedKey . '.o_id = ' . $table . '.o_id'
-                        . ' and ' . $mappedKey . '.fieldname = ' . $db->quote($fieldname)
-                        . ' and ' . $mappedKey . '.groupId=' . $featureJoin['groupId']
-                        . ' and ' . $mappedKey . '.keyId=' . $featureJoin['keyId']
-                        . ' and ' . $mappedKey . '.language = ' . $db->quote($featureJoin['language'])
-                        . ')',
-                        [
-                            $mappedKey => 'value'
-                        ]
-                    );
-                }
-
-                $havings = $featureFilters['conditions'];
-                if ($havings) {
-                    $havings = implode(' AND ', $havings);
-                    $select->having($havings);
-                }
-            });
-        }
     }
 
     /**
